+++ demo/src/main/java/com/example/demo/dao/ExchangeRepo.java
@@ -3,6 +3,6 @@ package com.example.demo.dao;
 import com.example.demo.dao.model.Exchange;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExchangeRepo extends JpaRepository<Exchange,Integer> {
+public interface ExchangeRepo extends JpaRepository<Exchange,Long> {
 
 }
